+++ 1.21/Fabric/src/main/java/fuzs/puzzleslib/fabric/mixin/server/MainFabricMixin.java
@@ -7,17 +7,15 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-/**
- * Run after Fabric Data Generation Api for same behavior as Forge where load complete does not run during data
- * generation (in case we ever decide to use Fabric's data generation api).
- */
-@Mixin(value = Main.class, priority = 1200)
+@Mixin(Main.class)
 abstract class MainFabricMixin {
 
     @Inject(
-            method = "main", at = @At(value = "NEW", target = "net/minecraft/server/dedicated/DedicatedServerSettings")
+            method = "main", at = @At(value = "NEW", target = "net/minecraft/server/dedicated/DedicatedServerSettings", shift = At.Shift.AFTER)
     )
     private static void main(String[] strings, CallbackInfo callback) {
-        FabricLifecycleEvents.SERVER_LOAD_COMPLETE.invoker().onLoadComplete();
+        // run after Fabric Data Generation Api for same behavior as Forge where load complete does not run
+        // during data generation (not that we use Fabric's data generation, but ¯\_(ツ)_/¯)
+        FabricLifecycleEvents.LOAD_COMPLETE.invoker().onLoadComplete();
     }
 }